+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+
+* Skroutz.Http.serialize_params to format a list|map as query params
+* Skroutz.Notification module mapping user notifications
+* Skroutz.Favorite module mapping user favorites
+* Skroutz.FavoriteList module mapping user favorite lists
+
+### Fixed
+
+* 204 No content response bodies no longer get parsed as JSON
+
 ## [0.0.1] - 2016-02-21
 
 ### Added